tests/wpt/web-platform-tests/app-history/navigate-event/transitionWhile-and-navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(r => window.onload = () => t.step_timeout(r, 0));
  await appHistory.navigate("#1").finished;

  appHistory.onnavigate = e => e.transitionWhile(Promise.resolve());
  appHistory.oncurrentchange = e => {
    if (e.navigationType == "traverse") {
      assert_equals(location.hash, "");
      assert_equals(appHistory.current.index, 0);
      assert_equals(appHistory.entries().length, 2);
      appHistory.navigate("#2");
    }
  }
  let back_result = appHistory.back();
  await promise_rejects_dom(t, "AbortError", back_result.committed);
  await promise_rejects_dom(t, "AbortError", back_result.finished);
  assert_equals(location.hash, "#2");
  assert_equals(appHistory.current.index, 1);
  assert_equals(appHistory.entries().length, 2);
}, "Using transitionWhile then navigate() in the ensuing currentchange should abort the transitionWhile");
</script>
